backend/requirements.txt
fastapi
pyyaml
openai
azure-identity
uvicorn
langchain-openai
